feat: move TranslationConfig to commons, fix cross-module smart casts
