+++ amethyst/src/play/java/com/vitorpamplona/amethyst/ui/components/TranslatableRichTextViewer.kt
@@ -140,14 +140,17 @@ private fun RenderTextWithTranslateOptions(
     Column {
         displayText(toBeViewed)
 
+        val sourceLang = translatedTextState.sourceLang
+        val targetLang = translatedTextState.targetLang
+
         if (
-            translatedTextState.sourceLang != null &&
-            translatedTextState.targetLang != null &&
-            translatedTextState.sourceLang != translatedTextState.targetLang
+            sourceLang != null &&
+            targetLang != null &&
+            sourceLang != targetLang
         ) {
             TranslationMessage(
-                translatedTextState.sourceLang,
-                translatedTextState.targetLang,
+                sourceLang,
+                targetLang,
                 translationMessageModifier,
                 accountViewModel,
             ) {
@@ -365,18 +368,21 @@ fun TranslateAndWatchLanguageChanges(
                     accountViewModel.translateTo(),
                 ).addOnCompleteListener { task ->
                     if (task.isSuccessful && !content.equals(task.result.result, true)) {
-                        if (task.result.sourceLang != null && task.result.targetLang != null) {
+                        val resultSourceLang = task.result.sourceLang
+                        val resultTargetLang = task.result.targetLang
+
+                        if (resultSourceLang != null && resultTargetLang != null) {
                             val preference =
                                 accountViewModel.account.settings.preferenceBetween(
-                                    task.result.sourceLang!!,
-                                    task.result.targetLang!!,
+                                    resultSourceLang,
+                                    resultTargetLang,
                                 )
                             val newConfig =
                                 TranslationConfig(
                                     result = task.result.result,
-                                    sourceLang = task.result.sourceLang,
-                                    targetLang = task.result.targetLang,
-                                    showOriginal = preference == task.result.sourceLang,
+                                    sourceLang = resultSourceLang,
+                                    targetLang = resultTargetLang,
+                                    showOriginal = preference == resultSourceLang,
                                 )
 
                             onTranslated(newConfig)
